hot fixes

+++ parse.go
@@ -41,8 +41,8 @@ func Strfduration(duration time.Duration, locale language.Tag) (string, error) {
 		fmt.Println("locale =", locale.String()+"_EG", " duration = ", int(duration.Minutes()), "arr", lc.Numbers, "transelated", lc.translateNumber(int(duration.Minutes())))
 		return lc.translateNumber(int(duration.Minutes())), nil
 	default:
-		fmt.Println("english duration = ", fmt.Sprintf("%.0v", duration.Minutes()))
-		return fmt.Sprintf("%.0v", duration.Minutes()), nil
+		fmt.Println("english duration = ", fmt.Sprintf("%.0v", int32(duration.Minutes())))
+		return fmt.Sprintf("%.0v", int32(duration.Minutes())), nil
 	}
 }
